+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/trie/pathbased/bonsai/worldview/BonsaiWorldState.java
@@ -266,9 +266,9 @@ public class BonsaiWorldState extends PathBasedWorldState {
       final PathBasedValue<BonsaiAccount> accountValue =
           worldStateUpdater.getAccountsToUpdate().get(updatedAddress);
       final BonsaiAccount accountOriginal = accountValue.getPrior();
+      final boolean storageCleared = worldStateUpdater.getStorageToClear().contains(updatedAddress);
       final Hash storageRoot =
-          (accountOriginal == null
-                  || worldStateUpdater.getStorageToClear().contains(updatedAddress))
+          (accountOriginal == null || storageCleared)
               ? Hash.EMPTY_TRIE_HASH
               : accountOriginal.getStorageRoot();
       final MerkleTrie<Bytes, Bytes> storageTrie =
@@ -285,7 +285,7 @@ public class BonsaiWorldState extends PathBasedWorldState {
         final UInt256 updatedStorage = storageUpdate.getValue().getUpdated();
         try {
 
-          if (!storageUpdate.getValue().isUnchanged()) {
+          if (storageCleared || !storageUpdate.getValue().isUnchanged()) {
             if (updatedStorage == null || updatedStorage.equals(UInt256.ZERO)) {
               maybeStateUpdater.ifPresent(
                   bonsaiUpdater ->